Move main loop to its own dir. Leave backwards compatibility entry point
